Here's a commit message for the changes made to the `requirements.txt` file:
```
Add whimsical rabbit-themed comment to requirements.txt

A playful comment about the importance of the requirements file, likening it to the essential role of rabbits in a magician's act, has been added to the top of the requirements.txt file. This serves to both entertain and remind users of the file's purpose.
```