Here's a commit message for the changes made to the README.md file:
```
Add whimsical rabbit-themed comment to README

In an effort to infuse a touch of charm and whimsy, a comment about rabbits has been added to the README.md file. This serves as a playful prelude to the serious topic of password generation, drawing a parallel between the enigmatic nature of rabbits and the world of cryptography.
```