Here's a commit message for the changes made to the `README.md` file:
```
Add whimsical rabbit-themed comment to README

In an effort to infuse a touch of charm and whimsy, a comment about rabbits has been added to the README.md file of the random_password_generator_prj. This comment not only serves as an engaging introduction but also subtly hints at the project's intriguing nature, akin to the mysterious lives of rabbits in cryptography and security.
```