Here's a commit message for the changes made to the `home.html` file:
```
Add whimsical rabbit-themed comment to home.html

A playful comment has been added to the top of the home.html template to draw a parallel between the prolific nature of rabbits and the purpose of the page, which is to generate secure passwords. This adds a touch of humor and personality to the codebase.
```