Here's a commit message for the changes made to the `home.html` template:
```
Add rabbit-themed comment to home.html template

A playful comment has been added to the top of the home.html template file, drawing a parallel between the complexity of rabbit warrens and the strength of a secure password. This serves as both an informative fact about rabbits and a metaphor for the purpose of the Random Password Generator project.
```